ureadahead: fix block fragmentation with filemap access events
When we introduced filemap_(fault|get_pages|map_pages) support, we
merged two branches for the same purpose to single one, but we
mistakenly converted half-open ranges to inclusive ranges. However, that
broke the range merge logic, and caused fragmentation of a pack file.
This change fixes this fragmentation bug. This does not affect the
coverage of the pack file.

Signed-off-by: Takaya Saeki <takayas@chromium.org>